generate_captions_1_2/non-docker/requirements.txt
transformers==4.15.0
timm==0.4.12
fairscale==0.4.4
torch
requests
Pillow